refactor(auto_search): extract search_target.rs (types + target builders)
Second slice of the Phase 1 split. Extracted:

  pub enum SearchTarget + impl SearchTarget::for_episode
  pub fn build_missing_targets
  pub fn build_monitored_targets
  pub fn build_upgrade_targets
  + 11 #[test] fns covering SearchTarget::for_episode + the `detail_with`
    test-helper they share.

Public API preserved via `pub use search_target::{SearchTarget,
build_missing_targets, build_monitored_targets, build_upgrade_targets};`
in mod.rs, so every external caller (main.rs, rss.rs, auto_expand.rs,
handlers) still resolves on unchanged paths.

Internal coupling with mod.rs is limited to one call:
`build_upgrade_targets` uses `resolve_existing_classification` which
stays pub in mod.rs; search_target.rs imports it as `use super::…`.

mod.rs: 6305 → 6006 lines. search_target.rs: 325 lines.
Tests still passing: 601.
